// src/models/enums.model.ts
export enum OS {
  WINDOWS = 'Windows',
  MAC = 'Mac',
  UBUNTU = 'Ubuntu',
  UNKNOWN = 'Unknown',
}

export enum BROWSER {
  CHROME = 'Chrome',
  SAFARI = 'Safari',
  FIREFOX = 'Firefox',
  POSTMAN = 'Postman',
  UNKNOWN = 'Unknown',
}

export enum LANGUAGE {
  ES = 'Spanish',
  EN = 'English',
  DE = 'German',
  PT = 'Portuguese',
  FR = 'French',
  UNKNOW = 'Unknow',
}